+++ Wrappers/Python/cil/processors/__init__.py
@@ -26,4 +26,5 @@ from .MaskGenerator import MaskGenerator
 from .AbsorptionTransmissionConverter import AbsorptionTransmissionConverter
 from .TransmissionAbsorptionConverter import TransmissionAbsorptionConverter
 from .Masker import Masker
+from .Padder import Padder